requirements.txt
asgiref==3.3.1
astpretty==2.1.0
atomicwrites==1.4.0
attrs==20.3.0
cognitive-complexity==1.2.0
colorama==0.4.4
Django==3.1.4
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-tinymce4-lite==1.8.0
eradicate==2.0.0
flake8==3.8.4
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-django==1.1.1
flake8-eradicate==1.0.0
flake8-expression-complexity==0.0.9
flake8-functions==0.0.5
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-variables-names==0.0.4
gunicorn==20.0.4
iniconfig==1.1.1
jsmin==2.2.2
mccabe==0.6.1
mr-proper==0.0.6
mypy-extensions==0.4.3
packaging==20.8
pep8-naming==0.11.1
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stdlib-list==0.8.0
toml==0.10.2
typing-extensions==3.7.4.3
